projects/sitemanage/src/main/java/com/percussion/widgetbuilder/utils/xform/PSContentTypeFileTransformer.java
/*
 *     Percussion CMS
 *     Copyright (C) 1999-2020 Percussion Software, Inc.
 *
 *     This program is free software: you can redistribute it and/or modify
 *     it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
 *
 *     This program is distributed in the hope that it will be useful,
 *     but WITHOUT ANY WARRANTY; without even the implied warranty of
 *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 *     GNU Affero General Public License for more details.
 *
 *     Mailing Address:
 *
 *      Percussion Software, Inc.
 *      PO Box 767
 *      Burlington, MA 01803, USA
 *      +01-781-438-9900
 *      support@percussion.com
 *      https://www.percusssion.com
 *
 *     You should have received a copy of the GNU Affero General Public License along with this program.  If not, see <https://www.gnu.org/licenses/>
 */
package com.percussion.widgetbuilder.utils.xform;

import com.percussion.cms.objectstore.PSItemDefinition;
import com.percussion.design.objectstore.PSApplyWhen;
import com.percussion.design.objectstore.PSBackEndColumn;
import com.percussion.design.objectstore.PSBackEndTable;
import com.percussion.design.objectstore.PSConditionalExit;
import com.percussion.design.objectstore.PSControlDependencyMap;
import com.percussion.design.objectstore.PSControlMeta;
import com.percussion.design.objectstore.PSControlRef;
import com.percussion.design.objectstore.PSDependency;
import com.percussion.design.objectstore.PSDisplayMapper;
import com.percussion.design.objectstore.PSDisplayMapping;
import com.percussion.design.objectstore.PSDisplayText;
import com.percussion.design.objectstore.PSExtensionCall;
import com.percussion.design.objectstore.PSExtensionCallSet;
import com.percussion.design.objectstore.PSExtensionParamValue;
import com.percussion.design.objectstore.PSField;
import com.percussion.design.objectstore.PSFieldSet;
import com.percussion.design.objectstore.PSFieldTranslation;
import com.percussion.design.objectstore.PSFieldValidationRules;
import com.percussion.design.objectstore.PSInputTranslations;
import com.percussion.design.objectstore.PSParam;
import com.percussion.design.objectstore.PSRule;
import com.percussion.design.objectstore.PSSearchProperties;
import com.percussion.design.objectstore.PSSingleHtmlParameter;
import com.percussion.design.objectstore.PSTableRef;
import com.percussion.design.objectstore.PSTableSet;
import com.percussion.design.objectstore.PSTextLiteral;
import com.percussion.design.objectstore.PSUISet;
import com.percussion.extension.PSExtensionRef;
import com.percussion.tablefactory.PSJdbcColumnDef;
import com.percussion.tablefactory.PSJdbcDataTypeMap;
import com.percussion.tablefactory.PSJdbcTableSchema;
import com.percussion.util.IOTools;
import com.percussion.util.PSCollection;
import com.percussion.widgetbuilder.data.PSWidgetBuilderFieldData;
import com.percussion.widgetbuilder.data.PSWidgetBuilderFieldData.FieldType;
import com.percussion.widgetbuilder.utils.IPSWidgetFileTransformer;
import com.percussion.widgetbuilder.utils.PSWidgetPackageBuilderException;
import com.percussion.widgetbuilder.utils.PSWidgetPackageSpec;
import com.percussion.xml.PSXmlDocumentBuilder;

import java.io.File;
import java.io.IOException;
import java.io.Reader;
import java.io.StringReader;
import java.io.StringWriter;
import java.io.Writer;
import java.sql.Types;
import java.util.ArrayList;
import java.util.HashMap;
import java.util.Iterator;
import java.util.List;
import java.util.Map;

import org.apache.commons.lang.Validate;
import org.w3c.dom.Element;
import org.xml.sax.SAXException;

/**
 * Transforms the files used to generate the content type
 * 
 * @author JaySeletz
 *
 */
public class PSContentTypeFileTransformer implements IPSWidgetFileTransformer {
	/**
     * 
     */
	private static final String MAX_TEXT_LEN = "255";
	private static final String FILE_PATH_LEN = "1000";
	private static final String IMG_PATH_LEN = "1000";
	private static final String PAGE_PATH_LEN = "1000";
	private static final String LINK_LEN = "15";
	
	private static final String SUFFIX = ".contentType";
	private static final String RTE_CONTROL = "sys_tinymce";
	private static final String REQUIRES_CLEANUP = "yes";

	private static PSJdbcDataTypeMap dataTypeMap;
	private static Map<String, Integer> dbColumnTypeMap;
	private static Map<String, String> controlTypeMap;

	private int nextObjectId = 1000;
	private IPSControlManager ctrlMgr;

	static {
		dbColumnTypeMap = new HashMap<String, Integer>();
		dbColumnTypeMap.put(FieldType.DATE.name(), Types.TIMESTAMP);
		dbColumnTypeMap.put(FieldType.RICH_TEXT.name(), Types.CLOB);
		dbColumnTypeMap.put(FieldType.TEXT.name(), Types.VARCHAR);
		dbColumnTypeMap.put(FieldType.TEXT_AREA.name(), Types.CLOB);
		dbColumnTypeMap.put(FieldType.FILE.name(), Types.VARCHAR);
		dbColumnTypeMap.put(FieldType.IMAGE.name(), Types.VARCHAR);
		dbColumnTypeMap.put(FieldType.PAGE.name(), Types.VARCHAR);
		dbColumnTypeMap.put(FieldType.IMAGE_LINK.name(), Types.INTEGER);
		dbColumnTypeMap.put(FieldType.PAGE_LINK.name(), Types.INTEGER);
		dbColumnTypeMap.put(FieldType.FILE_LINK.name(), Types.INTEGER);
		
		//TODO : FILE_LINK, IMAGE_LINK and PAGE_LINK should be Types.INTEGER
		// Need to make sure does not break on upgrade.
		
		controlTypeMap = new HashMap<String, String>();
		controlTypeMap.put(FieldType.DATE.name(), "sys_CalendarSimple");
		controlTypeMap.put(FieldType.RICH_TEXT.name(), RTE_CONTROL);
		controlTypeMap.put(FieldType.TEXT.name(), "sys_EditBox");
		controlTypeMap.put(FieldType.TEXT_AREA.name(), "sys_TextArea");
		controlTypeMap.put(FieldType.FILE.name(), "sys_FilePath");
		controlTypeMap.put(FieldType.FILE_LINK.name(), "sys_HiddenInput");
		controlTypeMap.put(FieldType.IMAGE.name(), "sys_ImagePath");
		controlTypeMap.put(FieldType.IMAGE_LINK.name(), "sys_HiddenInput");
		controlTypeMap.put(FieldType.PAGE.name(), "sys_PagePath");
        controlTypeMap.put(FieldType.PAGE_LINK.name(), "sys_HiddenInput");

	}

	public PSContentTypeFileTransformer(IPSControlManager ctrlMgr) {
		Validate.notNull(ctrlMgr);
		this.ctrlMgr = ctrlMgr;
	}

	@Override
	public Reader transformFile(File file, Reader reader,
			PSWidgetPackageSpec packageSpec)
			throws PSWidgetPackageBuilderException {
		try {
			Reader result = null;

			if (isSchemaFile(file))
				result = transformSchema(reader, packageSpec);
			else if (isItemDef(file))
				result = transformItemDef(reader, packageSpec);
			else
				result = reader;

			return result;
		} catch (Exception e) {
			throw new PSWidgetPackageBuilderException(
					"Failed to transform content type definition file: "
							+ file.getName(), e);
		}
	}

	private Reader transformSchema(Reader reader,
			PSWidgetPackageSpec packageSpec) throws Exception {
		PSJdbcTableSchema schema = getSchema(reader);
		List<PSWidgetBuilderFieldData> fields = packageSpec.getFields();
		for (PSWidgetBuilderFieldData field : fields) {
			schema.setColumn(new PSJdbcColumnDef(getDataTypeMap(), field
					.getName().toUpperCase(), PSJdbcColumnDef.ACTION_CREATE,
					getDbType(field), getSize(field), true, ""));
			if (field.getType().equals(FieldType.IMAGE.name())
			        || field.getType().equals(FieldType.FILE.name()) 
			        || field.getType().equals(FieldType.PAGE.name())) {
				schema.setColumn(new PSJdbcColumnDef(getDataTypeMap(), field
						.getName().toUpperCase() + "_LINKID",
						PSJdbcColumnDef.ACTION_CREATE, Types.INTEGER,
						getSize(field), true, ""));
			}
		}
		return new StringReader(PSXmlDocumentBuilder.toString(schema
				.toXml(PSXmlDocumentBuilder.createXmlDocument())));
	}

	private Reader transformItemDef(Reader reader,
			PSWidgetPackageSpec packageSpec) throws Exception {
		PSItemDefinition itemDef = getItemDef(reader);

		// workflow id is automatically updated, need to set token so that
		// descriptor uses the new id
		packageSpec.getResolverTokenMap().put("WORKFLOW_ID",
				String.valueOf(itemDef.getWorkflowId()));

		PSTableSet tableSet = itemDef.getTableSet();
		PSTableRef tableRef = (PSTableRef) tableSet.getTableRefs().next();
		PSBackEndTable beTable = new PSBackEndTable(tableRef.getAlias());
		PSFieldSet fieldSet = itemDef.getFieldSet();
		PSDisplayMapper mapper = itemDef.getDisplayMapper(fieldSet.getName());
		if (mapper == null) {
			// this is a bug, should always be found
			throw new RuntimeException(
					"No matching display mapper found for fieldset: "
							+ fieldSet.getName());
		}

		List<PSWidgetBuilderFieldData> fields = packageSpec.getFields();
		if (fields == null || fields.isEmpty())
			throw new RuntimeException(
					"Package spec must contain at least one field");

		boolean addedImageProcessors = false;
		for (PSWidgetBuilderFieldData field : fields) {
			//Widget Builder fields should not be required by default.
			PSField psfield = addField(beTable, field, fieldSet, false);
			PSDisplayMapping mapping = addMapping(field, mapper);

			if (field.getType().equals(FieldType.RICH_TEXT.name())) {
				// Add xdtextcleanup extension and control mapping for rich text
				// fields
				addTextCleanupExtension(mapping, itemDef);
				addReservedHtmlClassCleanerExtension(psfield, itemDef);
				addRichTextLinkFieldTranslations(psfield);
			} else if (field.getType().equals(FieldType.FILE.name())) {
				// add managed link processors/transformers
				if (!addedImageProcessors)
					addImageProcessors(itemDef);

				// add "_linkId" field and transforms needed for manage links
				PSWidgetBuilderFieldData linkIdField = new PSWidgetBuilderFieldData();
				linkIdField.setName(field.getName() + "_linkId");
				linkIdField.setLabel(linkIdField.getName());
				linkIdField.setType(FieldType.FILE_LINK.name());
				PSField psLinkField = addField(beTable, linkIdField, fieldSet,
						false);
				addImgLinkFieldTranslations(psfield, psLinkField);
				addMapping(linkIdField, mapper);

				
			} else if (field.getType().equals(FieldType.IMAGE.name())) {
				// add managed link processors/transformers
				if (!addedImageProcessors)
					addImageProcessors(itemDef);
				
				// add "_linkId" field and transforms needed for manage links
				PSWidgetBuilderFieldData linkIdField = new PSWidgetBuilderFieldData();
				linkIdField.setName(field.getName() + "_linkId");
				linkIdField.setLabel(linkIdField.getName());
				linkIdField.setType(FieldType.IMAGE_LINK.name());
		
				PSField psLinkField = addField(beTable, linkIdField, fieldSet,
						false);
				addImgLinkFieldTranslations(psfield, psLinkField);
				addMapping(linkIdField, mapper);

			} else if (field.getType().equals(FieldType.PAGE.name())) {
                // add managed link processors/transformers
                if (!addedImageProcessors)
                    addImageProcessors(itemDef);
                
                // add "_linkId" field and transforms needed for manage links
                PSWidgetBuilderFieldData linkIdField = new PSWidgetBuilderFieldData();
                linkIdField.setName(field.getName() + "_linkId");
                linkIdField.setLabel(linkIdField.getName());
                linkIdField.setType(FieldType.IMAGE_LINK.name());
                PSField psLinkField = addField(beTable, linkIdField, fieldSet,
                        false);
                addImgLinkFieldTranslations(psfield, psLinkField);
                addMapping(linkIdField, mapper);

            }
		}

		return new StringReader(PSXmlDocumentBuilder.toString(itemDef
				.toXml(PSXmlDocumentBuilder.createXmlDocument())));
	}

	private void addReservedHtmlClassCleanerExtension(PSField field,
			PSItemDefinition itemDef) {
		PSInputTranslations inputTranslations = new PSInputTranslations();
		Iterator<?> currentTrans = itemDef.getContentEditor()
				.getInputTranslations();
		while (currentTrans.hasNext()) {
			inputTranslations.add(currentTrans.next());
		}

		PSExtensionCallSet callSet = new PSExtensionCallSet();
		PSExtensionParamValue[] params = new PSExtensionParamValue[1];
		params[0] = new PSExtensionParamValue(new PSTextLiteral(
				field.getSubmitName()));
		callSet.add(new PSExtensionCall(new PSExtensionRef(
				"Java/global/percussion/content/sys_cleanReservedHtmlClasses"),
				params));
		inputTranslations.add(new PSConditionalExit(callSet));
		itemDef.getContentEditor().setInputTranslation(inputTranslations);
	}

	private void addImgLinkFieldTranslations(PSField imgField, PSField linkField) {
		PSExtensionCallSet callSet = new PSExtensionCallSet();
		PSExtensionRef ref = new PSExtensionRef(
				"Java/global/percussion/content/sys_manageItemPathOnUpdate");
		PSExtensionParamValue[] params = new PSExtensionParamValue[2];
		params[0] = new PSExtensionParamValue(new PSSingleHtmlParameter(
				imgField.getSubmitName()));
		params[1] = new PSExtensionParamValue(new PSSingleHtmlParameter(
				linkField.getSubmitName()));
		callSet.add(new PSExtensionCall(ref, params));
		PSFieldTranslation fieldTranslation = new PSFieldTranslation(callSet);
		linkField.setInputTranslation(fieldTranslation);

		callSet = new PSExtensionCallSet();
		ref = new PSExtensionRef(
				"Java/global/percussion/content/sys_manageItemPathOnEdit");
		params = new PSExtensionParamValue[2];
		params[0] = new PSExtensionParamValue(new PSSingleHtmlParameter(
				imgField.getSubmitName()));
		params[1] = new PSExtensionParamValue(new PSTextLiteral(
				linkField.getSubmitName()));
		callSet.add(new PSExtensionCall(ref, params));
		fieldTranslation = new PSFieldTranslation(callSet);
		imgField.setOutputTranslation(fieldTranslation);
	}

	private PSField addField(PSBackEndTable beTable,
			PSWidgetBuilderFieldData field, PSFieldSet fieldSet,
			boolean required) throws Exception {
		// create locator
		PSBackEndColumn col = new PSBackEndColumn(beTable, field.getName()
				.toUpperCase());

		// add field to field set
		PSField psfield = new PSField(field.getName(), col);
		psfield.setType(PSField.TYPE_LOCAL);
		psfield.setMimeType("text/plain");
		setTypeSpecificFieldProperties(field, psfield);
		psfield.setSearchProperties(new PSSearchProperties(true));
		psfield.setOccurrenceDimension(PSField.OCCURRENCE_DIMENSION_OPTIONAL,
				null);
		if (required)
			addValidationRule(psfield, field);
		fieldSet.add(psfield);

		return psfield;
	}

	private PSDisplayMapping addMapping(PSWidgetBuilderFieldData field,
			PSDisplayMapper mapper) throws Exception {
		// add mapping to ui set
		PSUISet uiSet = new PSUISet();
		uiSet.setLabel(new PSDisplayText(field.getLabel() + ":"));
		uiSet.setErrorLabel(new PSDisplayText(field.getLabel() + ":"));
		uiSet.setControl(getControlRef(field));

		PSDisplayMapping mapping = new PSDisplayMapping(field.getName(), uiSet);
		mapper.add(mapping);
		return mapping;
	}

	/**
	 * Add necessary input transform and result doc processor for image fields
	 * 
	 * @param itemDef
	 *            The itemdef to update, not <code>null</code>.
	 */
	private void addImageProcessors(PSItemDefinition itemDef) {
		PSInputTranslations inputTranslations = new PSInputTranslations();

		Iterator<?> currentTrans = itemDef.getContentEditor()
				.getInputTranslations();
		while (currentTrans.hasNext()) {
			inputTranslations.add(currentTrans.next());
		}

		PSExtensionCallSet callSet = new PSExtensionCallSet();
		callSet.add(new PSExtensionCall(
				new PSExtensionRef(
						"Java/global/percussion/content/sys_managedItemPathPreProcessor"),
				null));
		inputTranslations.add(new PSConditionalExit(callSet));
		itemDef.getContentEditor().setInputTranslation(inputTranslations);

		callSet = new PSExtensionCallSet();
		callSet.add(new PSExtensionCall(new PSExtensionRef(
				"Java/global/percussion/content/sys_manageLinksPostProcessor"),
				null));
		itemDef.getPipe().setResultDataExtensions(callSet);
	}

	private void setTypeSpecificFieldProperties(PSWidgetBuilderFieldData field,
			PSField psfield) {
		if (field.getType().equals(FieldType.RICH_TEXT.name())) {
			psfield.setAllowActiveTags(false);
			psfield.setCleanupBrokenInlineLinks(true);
			psfield.setCleanupNamespaces(true);
			psfield.setMayHaveInlineLinks(true);
			psfield.setDataFormat("max");
			psfield.setDataType(PSField.DT_TEXT);
		} else if (field.getType().equals(FieldType.DATE.name())) {
			psfield.setDataType(PSField.DT_DATE);
		} else if (field.getType().equals(FieldType.FILE.name())) {
			psfield.setDataType(PSField.DT_BINARY);
		} else if (field.getType().equals(FieldType.TEXT_AREA.name())) {
			psfield.setDataType(PSField.DT_TEXT);
			psfield.setDataFormat("max");
		} else if(field.getType().equals(FieldType.PAGE_LINK.name()) ||
				field.getType().equals(FieldType.FILE_LINK.name()) ||
				field.getType().equals(FieldType.IMAGE_LINK.name())){
		      psfield.setDataType(PSField.DT_INTEGER);
		}else {
			psfield.setDataType("text");
			psfield.setDataFormat(getSize(field));
		}
	}

	private PSControlRef getControlRef(PSWidgetBuilderFieldData field) {
		PSControlRef controlRef = new PSControlRef(controlTypeMap.get(field
				.getType()));
		controlRef.setId(getNextObjectId());
		if (field.getType().equals(FieldType.TEXT.name())) {
			PSCollection params = new PSCollection(PSParam.class);
			params.add(new PSParam("maxlength", new PSTextLiteral(MAX_TEXT_LEN)));
			controlRef.setParameters(params);
		} else if (field.getType().equals(FieldType.TEXT_AREA.name())
				|| field.getType().equals(FieldType.RICH_TEXT.name())) {
			PSCollection params = new PSCollection(PSParam.class);
			params.add(new PSParam("requirescleanup", new PSTextLiteral(
					REQUIRES_CLEANUP)));
			controlRef.setParameters(params);
		}

		return controlRef;
	}

	/**
	 * Adds the sys_xdTextCleanup extension to the input data exits, and a
	 * corresponding set of control dependency user properties
	 * 
	 * @param field
	 *            The field to add it for, not <code>null</code>
	 * @param itemDef
	 *            The item def to add to, not <code>null</code>.
	 */
	@SuppressWarnings("unchecked")
	private void addTextCleanupExtension(PSDisplayMapping mapping,
			PSItemDefinition itemDef) {
		// add extension, add to map
		PSControlDependencyMap depMap = itemDef.getPipe()
				.getControlDependencyMap();
		PSControlMeta ctrlMeta = ctrlMgr.getControl(mapping.getUISet()
				.getControl().getName());
		List<PSDependency> deps = new ArrayList<PSDependency>();
		deps.addAll(ctrlMeta.getDependencies());
		for (PSDependency dep : deps) {
			dep.setId(getNextObjectId());
		}
		depMap.setControlDependencies(mapping, deps);
	}

	/**
	 * Adds manage link converter extension to rich text field.
	 * 
	 * @param psField
	 *            assumed not <code>null</code>
	 */
	private void addRichTextLinkFieldTranslations(PSField psField) {
		PSExtensionCallSet callSet = new PSExtensionCallSet();
		PSExtensionRef ref = new PSExtensionRef(
				"Java/global/percussion/content/sys_manageLinksConverter");
		PSExtensionParamValue[] params = new PSExtensionParamValue[1];
		params[0] = new PSExtensionParamValue(new PSSingleHtmlParameter(
				psField.getSubmitName()));
		callSet.add(new PSExtensionCall(ref, params));
		PSFieldTranslation fieldTranslation = new PSFieldTranslation(callSet);
		psField.setInputTranslation(fieldTranslation);
	}

	/**
	 * @param psfield
	 * @param field
	 */
	private void addValidationRule(PSField psfield,
			PSWidgetBuilderFieldData field) {
		PSFieldValidationRules validationRules = new PSFieldValidationRules();

		PSCollection rules = new PSCollection(PSRule.class);
		PSExtensionCallSet extensions = new PSExtensionCallSet();
		PSExtensionParamValue[] params = { new PSExtensionParamValue(
				new PSSingleHtmlParameter(field.getName())) };
		PSExtensionCall extension = new PSExtensionCall(new PSExtensionRef(
				"Java/global/percussion/content/sys_ValidateRequiredField"),
				params);
		extensions.add(extension);
		rules.add(new PSRule(extensions));
		validationRules.setRules(rules);

		PSApplyWhen applyWhen = new PSApplyWhen();
		applyWhen.setIfFieldEmpty(true);
		validationRules.setApplyWhen(applyWhen);

		String errorMsg = field.getLabel() + " may not be empty.";
		validationRules.setErrorMessage(new PSDisplayText(errorMsg));

		psfield.setValidationRules(validationRules);

	}

	private PSItemDefinition getItemDef(Reader reader) throws Exception {
		return new PSItemDefinition(getElementFromReader(reader));
	}

	/**
	 * Parse the content from the supplied reader into a Document and return the
	 * root element.
	 * 
	 * @param reader
	 *            The reader to use, does not close the underlying stream
	 * 
	 * @return The element, may be <code>null</code>.
	 * 
	 * @throws IOException
	 * @throws SAXException
	 */
	private Element getElementFromReader(Reader reader) throws IOException,
			SAXException {
		Writer out = new StringWriter();
		IOTools.writeStream(reader, out);
		return PSXmlDocumentBuilder.createXmlDocument(
				new StringReader(out.toString()), false).getDocumentElement();
	}

	private boolean isSchemaFile(File file) {
		return file.getName().endsWith(".schemaDef" + SUFFIX);
	}

	private boolean isItemDef(File file) {
		return file.getName().endsWith(".itemDef" + SUFFIX);
	}

	private String getSize(PSWidgetBuilderFieldData field) {
		if (FieldType.TEXT.name().equals(field.getType()))
			return MAX_TEXT_LEN;
		else if (FieldType.FILE.name().equals(field.getType()))
			return FILE_PATH_LEN;
		else if (FieldType.IMAGE.name().equals(field.getType()))
			return IMG_PATH_LEN;
		else if (FieldType.PAGE.name().equals(field.getType()))
            return PAGE_PATH_LEN;
		/*  Should set size and type of _link fields,  tests need resolving
		else if (FieldType.FILE_LINK.name().equals(field.getType()))
            return LINK_LEN;
		else if (FieldType.IMAGE_LINK.name().equals(field.getType()))
            return LINK_LEN;
		else if (FieldType.PAGE_LINK.name().equals(field.getType()))
            return LINK_LEN;
        */
		return null;
	}

	private int getDbType(PSWidgetBuilderFieldData field) {
		return dbColumnTypeMap.get(field.getType());
	}

	PSJdbcTableSchema getSchema(Reader reader) throws Exception {
		return new PSJdbcTableSchema(getElementFromReader(reader),
				getDataTypeMap());
	}

	private PSJdbcDataTypeMap getDataTypeMap() throws Exception {
		if (dataTypeMap == null)
			dataTypeMap = new PSJdbcDataTypeMap("DERBY", "", "");

		return dataTypeMap;
	}

	@Override
	public boolean handleFile(File file) {
		return isSchemaFile(file) || isItemDef(file);
	}

	@SuppressWarnings("unused")
	@Override
	public File transformPath(File file, PSWidgetPackageSpec packageSpec)
			throws PSWidgetPackageBuilderException {
		// no-op
		return file;
	}

	private int getNextObjectId() {
		return nextObjectId++;
	}

}
